Moved common mock into its own package for multiple test files to use
